Progress, but also a bit of a blind alley
Was about to embark upon a whole new mixin approach, but not needed.
Accordingly, dual_operators_mixin is about to depart this Earth.
dual_helpers is back, for now, but that was motivated by something I'm
no longer contemplating.  Bottom line, we're a bit all over the place
right now.

Key point.  Had to disable .value as that was what was making astropy
be greedy.  Workaround, if you want the value just do nd.array() of it
or units.Quantity().value?